+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.12.4
+
+* `tags_factory` in `trace_function` can be also a dict